backend/requirements.txt
fastapi[standard]
supabase
python-dotenv
pydantic
